@@ -1,9 +1,9 @@
 {
     "require": {
-        "php": "^8.0.2",
-        "phpstan/phpstan": "1.4.2",
-        "phpstan/phpstan-deprecation-rules": "1.0.0",
-        "phpstan/phpstan-strict-rules": "1.1.0",
+        "php": "^8.1",
+        "phpstan/phpstan": "1.10.3",
+        "phpstan/phpstan-deprecation-rules": "1.1.2",
+        "phpstan/phpstan-strict-rules": "1.5.0",
         "thecodingmachine/phpstan-strict-rules": "1.0.0",
         "ergebnis/phpstan-rules": "1.0.0"
     },
