@@ -103,10 +103,11 @@ abstract class AbstractApi
      * @param array<string,mixed>  $params
      * @param array<string,string> $headers
      * @param array<string,string> $files
+     * @param array<string,mixed>  $uriParams
      *
      * @return mixed
      */
-    protected function post(string $uri, array $params = [], array $headers = [], array $files = [])
+    protected function post(string $uri, array $params = [], array $headers = [], array $files = [], array $uriParams = [])
     {
         if (0 < \count($files)) {
             $builder = $this->createMultipartStreamBuilder($params, $files);
@@ -120,7 +121,7 @@ abstract class AbstractApi
             }
         }
 
-        $response = $this->client->getHttpClient()->post(self::prepareUri($uri), $headers, $body);
+        $response = $this->client->getHttpClient()->post(self::prepareUri($uri, $uriParams), $headers, $body);
 
         return ResponseMediator::getContent($response);
     }
