@@ -380,7 +380,7 @@ class RepositoriesTest extends TestCase
         $this->assertEquals($expectedArray, $api->commitRefs(1, 'abcd1234', ['type' => $type]));
     }
 
-    public function dataGetCommitRefsWithParams()
+    public static function dataGetCommitRefsWithParams(): array
     {
         return [
             'type_tag' => [
