@@ -140,7 +140,7 @@ class TagsTest extends TestCase
         $this->assertEquals($expectedResult, $api->updateRelease(1, $releaseName, $params));
     }
 
-    public function releaseDataProvider()
+    public static function releaseDataProvider(): array
     {
         return [
             [
